+++ ChangeLog.md
@@ -2,6 +2,12 @@
 
 ## REVISION HISTORY
 
+
+---
+- 2018-08-22 Version 6.1.0
+  - Implemented support for calling from parallel region, i.e. thread safe PMlib, for Intel compiler and PGI compiler.
+  - New API mergeThreads()/f_pm_mergethreads() to merge stats from task private threads
+
 ---
 - 2018-07-09 Version 6.0.0
   - Add support for thread level performance report